#86 Move shared confirm dialog
