+++ src/app/shared/confirm-dialog/confirm-dialog.component.html
@@ -3,15 +3,11 @@
     {{ message }}
   </mat-dialog-content>
   <mat-dialog-actions>
-    <button
-      (click)="onCancelClick()"
-      class="btn btn-secondary"
-      mat-raised-button>
+    <button (click)="onCancelClick()" class="btn-secondary" mat-raised-button>
       Abbrechen
     </button>
     <button
       (click)="onConfirmClick()"
-      class="btn"
       color="{{ actionType == 'Löschen' ? 'warn' : 'accent' }}"
       mat-raised-button>
       {{ actionType }}
